refactor: Service interface와 ServiceImpl구현체 분리 (#28)
